[UPD]迁移文件优化

+++ database/migrations/2021_08_27_141734_create_advert_users.php
@@ -22,7 +22,7 @@ class CreateAdvertUsers extends Migration
             $table->tinyInteger('type')->default(0)->comment('类型：1 幸运抽奖；2 答题；3 猜成语；4 刮刮乐；10 拆红包；11 拆红包翻倍；12 签到；13 签到翻倍；14 任务奖励；15 任务翻倍奖励；16 蘑菇；17 集碎片；18 达标兑换；19 步数兑换；20 气泡兑换；21 拉新；22 拉活');
             $table->tinyInteger('status')->default(1)->comment('状态：0 异常；1 正常');
             $table->tinyInteger('brand')->default(1)->comment('观看渠道：1 正常广告；2 拼团广告');
-            $table->unsignedInteger('unique_id')->default(0)->comment('唯一标识');
+            $table->unsignedBigInteger('unique_id')->default(0)->comment('唯一标识');
             $table->timestamp('created_at')->nullable()->comment('创建时间');
             $table->timestamp('updated_at')->nullable()->comment('更新时间');
